requirements.txt
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.3
idna==2.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pymongo==3.7.2
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
